Removed circular dependency of AxeOptions in background
