+++ src/scanner/launcher.ts
@@ -6,18 +6,6 @@ import { AxeResponseHandler } from './axe-response-handler';
 import { ScanOptions } from './scan-options';
 import { ScanParameterGenerator } from './scan-parameter-generator';
 
-export interface AxeOptions {
-    runOnly?: Axe.RunOnly;
-    restoreScroll?: Boolean; // missing from axe options.
-}
-
-export type AxeScanContext = string | NodeSelector & Node | IncludeExcludeOptions | NodeList;
-
-export interface IncludeExcludeOptions {
-    include: string[][];
-    exclude: string[][];
-}
-
 export class Launcher {
     constructor(
         private axe: typeof Axe,
