+++ src/tests/unit/tests/scanner/scan-parameter-generator.test.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 import { IMock, Mock, MockBehavior } from 'typemoq';
 
-import { AxeOptions } from '../../../../scanner/launcher';
+import { AxeOptions } from '../../../../scanner/axe-options';
 import { RuleSifter, RuleWithA11YCriteria } from '../../../../scanner/rule-sifter';
 import { ScanOptions } from '../../../../scanner/scan-options';
 import { ScanParameterGenerator } from '../../../../scanner/scan-parameter-generator';
